+++ bukkit/src/main/java/com/alessiodp/parties/bukkit/players/objects/ExpDrop.java
@@ -10,6 +10,6 @@ import org.bukkit.entity.Entity;
 public class ExpDrop {
 	@Getter private final PartyPlayerImpl killer;
 	@Getter private final Entity entityKilled;
-	@Getter @Setter private int normal;
-	@Getter @Setter private int skillApi;
+	@Getter @Setter private double normal;
+	@Getter @Setter private double skillApi;
 }
